requirements.txt
python-dotenv==1.0.0
cryptography==41.0.3
pywin32==306
pystray==0.19.4
Pillow==10.0.0
pysmb==1.2.9
psutil==5.9.5
packaging==23.1
schedule==1.2.0
loguru==0.7.0
apscheduler==3.10.1
pyinstaller==6.1.0
PyQt5==5.15.9 